Fix crash when updating libraries with multiple manifests
If a library has both a pyproject.toml file with PEP621 dependencies,
and a standard requirements.txt file, we'd end up using the `:widen`
strategy for the dependencies in the `requirements.txt` file and
eventually crashing with an error like the following:

```
/home/dependabot/dependabot-core/common/lib/dependabot/update_checkers/base.rb:266:in `block in preferred_version_resolvable_with_unlock?': undefined method `[]' for nil:NilClass (NoMethodError)

        updated_requirements.none? { |r| r[:requirement] == :unfixable }
                                          ^^^^^^^^^^^^^^
	from /home/dependabot/dependabot-core/common/lib/dependabot/update_checkers/base.rb:266:in `none?'
	from /home/dependabot/dependabot-core/common/lib/dependabot/update_checkers/base.rb:266:in `preferred_version_resolvable_with_unlock?'
	from /home/dependabot/dependabot-core/common/lib/dependabot/update_checkers/base.rb:249:in `numeric_version_can_update?'
	from /home/dependabot/dependabot-core/common/lib/dependabot/update_checkers/base.rb:199:in `version_can_update?'
	from /home/dependabot/dependabot-core/common/lib/dependabot/update_checkers/base.rb:44:in `can_update?'
	from bin/dry-run.rb:709:in `block in <main>'
	from bin/dry-run.rb:661:in `each'
	from bin/dry-run.rb:661:in `<main>'
```

This was introduced by the previous commit adding support for updating
libraries with `pyproject.toml` files using the `:widen` strategy, but
that was only intended for the dependencies in that file, not for other
dependencies, since that would mean a change over the existing behavior
(and one that does not work well).

So change the check to not only guard for existence of a
`pyproject.toml` file, but for actually updating a dependency inside
that file.